Revert "[ENG-1662] feat: Add provider initialization (#1205)"
This reverts commit 79dbd289aaa2c7ddba98375110cb3385097c4fa6.
