+++ connector/string.go
@@ -1,4 +1,4 @@
-package generic
+package connector
 
 func (c *Connector) String() string {
 	return c.Provider() + ".Connector"
